// src/app/interface/planPastoral.interface.ts
// Generated by https://quicktype.io

export interface ResultPlanPastorales {
  ok:           boolean;
  msg:          string;
  planpastoral: Planpastoral[];
}
export interface ResultPlanPastoralID {
  ok:           boolean;
  msg:          string;
  planpastoral: Planpastoral;
}
export interface Planpastoral {
  id:       number;
  titulo:   string;
  pdf:      string;
  posicion: number;
}

export interface PlanPastoralForm{
  titulo:string,
  pdf?:string,
  posicion:string
}
export interface PdfPlanPastoralForm{
  pdf?:string
}